Testing_software/testbench-website-streamlit/requirements.txt
python-dotenv~=1.0.0
mysql-connector-python~=8.3.0
tiktoken~=0.5.2
openai~=1.12.0
pandas~=2.1.4
streamlit~=1.31.1
langchain~=0.1.13
langchain-openai~=0.1.1
neo4j~=5.16.0
numpy~=1.26.3
pymupdf~=1.24.0
python-docx~=1.1.0